Move all tests to a tests folder
